Repositories' names changed and modify attribute Usuario
Cambio de nombres de las interfaces Repository poniendo el mismo nombre que la entidad, cambiando tambien el id de todas las entidades a tipo Long.
Introducir un atributo contraseña para cada usuario.

+++ BookTeca/src/main/java/dad/web/bookteca/clases/EquipoInformatico.java
@@ -8,13 +8,15 @@ public class EquipoInformatico {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private long id;
+	
 	private String sistemaOperativo;
 	private String localizacion;
-	
 	private boolean disponible;
+	
 	@OneToOne(cascade=CascadeType.ALL)
 	private Usuario idUsuario;
+	
 	private Date fechaReserva;
 	
 	public EquipoInformatico(String sistemaOperativo, String localizacion) {
@@ -22,7 +24,7 @@ public class EquipoInformatico {
 		this.localizacion = localizacion;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
